+++ amulet_map_editor/opengl/mesh/base/tri_mesh.py
@@ -115,7 +115,7 @@ class TriMesh(Drawable):
 
     def _draw(self, transformation_matrix: numpy.ndarray):
         glUseProgram(self._shader)
-        glUniformMatrix4fv(self._transform_location, 1, GL_FALSE, transformation_matrix)
+        glUniformMatrix4fv(self._transform_location, 1, GL_FALSE, transformation_matrix.astype(numpy.float32))
         glUniform1i(self._texture_location, 0)
         try:
             glBindVertexArray(self._vao)
